add simulator

+++ refactor/service/aircraft/impl/aircraft_service.cpp
@@ -95,7 +95,7 @@ void hako::service::impl::AircraftService::resetService()
 
 bool hako::service::impl::AircraftService::write_pdu(uint32_t index, HakoniwaPduDataType& pdu)
 {
-    if (index < 0 || index >= aircraft_inputs_.size()) {
+    if (index < 0 || index >= static_cast<int>(aircraft_inputs_.size())) {
         return false;
     }
 
@@ -127,7 +127,7 @@ bool hako::service::impl::AircraftService::write_pdu(uint32_t index, HakoniwaPdu
 
 bool hako::service::impl::AircraftService::read_pdu(uint32_t index, HakoniwaPduDataType& pdu)
 {
-    if (index < 0 || index >= aircraft_inputs_.size()) {
+    if (index < 0 || index >= static_cast<int>(aircraft_inputs_.size())) {
         return false;
     }
     switch (pdu.id) {
